SystemPerfApp/backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psutil>=5.9.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.3.0
